+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-09-24
+
+- Make RBGL an optional backend loaded with `termvas/rbgl`.
+- Use Tessel's shared quantizer for median-cut Sixel output.
+
 ## [0.1.0] - 2026-09-23
 
 - Initial release.